Add register redirect/check user id in edit + delete

+++ app.py
@@ -1,6 +1,6 @@
 import sqlite3
 from flask import Flask
-from flask import redirect, render_template, request, session
+from flask import abort, redirect, render_template, request, session
 from werkzeug.security import check_password_hash, generate_password_hash
 import config
 import db
@@ -49,15 +49,20 @@ def create_item():
 @app.route("/edit_item/<int:item_id>")
 def edit_item(item_id):
     item = items.get_item(item_id)
+    if item["user_id"] != session["user_id"]:
+        abort(403)
     return render_template("edit_item.html", item=item)
 
 @app.route("/update_item", methods=["POST"])
 def update_item():
+    item_id = request.form["item_id"]
+    item = items.get_item(item_id)
+    if item["user_id"] != session["user_id"]:
+        abort(403)
     book_name = request.form["book_name"]
     author = request.form["author"]
     grade = request.form["grade"]
     review = request.form["review"]
-    item_id = request.form["item_id"]
 
     items.update_item(item_id, book_name, author, grade, review)
 
@@ -65,8 +70,11 @@ def update_item():
 
 @app.route("/delete_item/<int:item_id>", methods = ["GET", "POST"])
 def delete_item(item_id):
+    item = items.get_item(item_id)
+    if item["user_id"] != session["user_id"]:
+        abort(403)
+
     if request.method == "GET":
-        item = items.get_item(item_id)
         return render_template("delete_item.html", item=item)
 
     if request.method == "POST":
@@ -95,7 +103,7 @@ def create():
     except sqlite3.IntegrityError:
         return "VIRHE: tunnus on jo varattu"
 
-    return "Tunnus luotu"
+    return redirect("/login")
 
 @app.route("/login", methods=["GET","POST"])
 def login():
